Rename admin users editor handler to users handler
